// homepage/src/modules/grpc/proto/v2.pb.ts
export interface Banner {
  position: number;
  id: string;
  bannerName: string;
  bannerImage: string;
  bannerPage: string;
  bannerPosition: string;
  bannerType: string;
  bannerValue: string;
  lang: string;
  createdAt: string;
  updatedAt: string;
}

export interface GetBannersResponse {
  banners: Banner[];
}

export interface GetBannersRequest {
  bannerPage: string[];
  bannerPosition: string;
  region: string;
  lang: string;
  type?: string | undefined;
}

export interface GetFeedRequest {
  size: number;
  feedTypes: string[];
  brands: string[];
  models: string[];
  categories: string[];
  category?: string | undefined;
}